refactor: post 패키지 코드 리팩터링
- 일부 구현된 facade 제거
- impl 제거 후 전체 코드 형식 통일
