+++ src/main/java/aibe/hosik/profile/ProfileService.java
@@ -2,7 +2,6 @@ package aibe.hosik.profile;
 
 import aibe.hosik.handler.exception.CustomException;
 import aibe.hosik.handler.exception.ErrorCode;
-import aibe.hosik.post.service.StorageService;
 import aibe.hosik.profile.dto.ProfileRequest;
 import aibe.hosik.profile.dto.ProfileResponse;
 import lombok.RequiredArgsConstructor;
